Sub list creation

+++ db/seeds.rb
@@ -5,15 +5,17 @@
 #
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
-MediaType.delete_all
-MediaType.create!(name:'Book')
-MediaType.create!(name:'Music')
-MediaType.create!(name:'Movie')
-MediaType.create!(name:'Course')
+unless Rails.env.test?
+  MediaType.delete_all
+  MediaType.create!(name:'Book')
+  MediaType.create!(name:'Music')
+  MediaType.create!(name:'Movie')
+  MediaType.create!(name:'Course')
 
-Language.delete_all
-Language.create!(name: 'French', code:'fr')
-Language.create!(name: 'English', code:'gb')
-Language.create!(name: 'Italian', code:'it')
-Language.create!(name: 'Spanish', code:'es')
-Language.create!(name: 'German', code:'al')
+  Language.delete_all
+  Language.create!(name: 'French', code:'fr')
+  Language.create!(name: 'English', code:'gb')
+  Language.create!(name: 'Italian', code:'it')
+  Language.create!(name: 'Spanish', code:'es')
+  Language.create!(name: 'German', code:'al')
+end
